databases/requirements.txt
psycopg2-binary==2.9.9
neo4j==5.11.0
numpy==1.26.2
pandas==2.1.3
sentence-transformers==2.2.2
tqdm==4.66.1
pgvector==0.2.3
requests==2.31.0
chromadb==0.4.22 